refactor(api): Move sitemap logic into seperate file
